crypto: Add wrapping primitives
This commit adds w8, w16, w32, w64 types which are re-export of
`std::num::wrapping::Wrapping`.

This commit also adds `ToWrapping` trait which provides methods for easy
conversion between wrapping types.

Note: due to overflow semantic changes, `as` operator is expected to
check overflow, although it is not implemented right now.
For future-proof, this implementation explicitly truncate before `as`
operation. When overflow semantics are fully implemented, it can be
replaced with `WrappingOps::wrapping_as_u32()` and so on.